# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Doc API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/suppliers": {
            "get": {
                "tags": [
                    "Supplier"
                ],
                "summary": "Список",
                "operationId": "9e0700fc31eb240038c790aa63361acb",
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Supplier"
                ],
                "summary": "Создание",
                "operationId": "930aa95e942c98b1828174a19b0ea497",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "ЧДММ Алюминчи"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Худжанд, 3мкр"
                                    },
                                    "phone": {
                                        "type": "integer",
                                        "example": 80800
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "alumin@mail.ru"
                                    },
                                    "status": {
                                        "type": "boolean",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/suppliers/{supplier}": {
            "get": {
                "tags": [
                    "Supplier"
                ],
                "summary": "Список по ID",
                "operationId": "ed59705564f1f6ae5e171db9c34d8c99",
                "parameters": [
                    {
                        "name": "supplier",
                        "in": "path",
                        "description": "ID клиент",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Supplier"
                ],
                "summary": "Удаление",
                "operationId": "fc72d1b5afa59f4ec61c7ffd30e13e7a",
                "parameters": [
                    {
                        "name": "supplier",
                        "in": "path",
                        "description": "ID клиент",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "done"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Supplier"
                ],
                "summary": "Обновление",
                "operationId": "65e598c4e07dbe419369d58f1bd4c2ce",
                "parameters": [
                    {
                        "name": "supplier",
                        "in": "path",
                        "description": "ID клиент",
                        "required": true,
                        "example": 2
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "ЧДММ Алюминчи"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Худжанд, 3мкр"
                                    },
                                    "phone": {
                                        "type": "integer",
                                        "example": 80800
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "alumin@mail.ru"
                                    },
                                    "status": {
                                        "type": "boolean",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/suppliers/filterByname": {
            "post": {
                "tags": [
                    "Supplier"
                ],
                "summary": "Список поставщиков с фильтрацией по название",
                "operationId": "3e419ebb8ae818ae13ad1e86bfb3b19d",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Название",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "ЧДММ"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список заказов, удовлетворяющих условиям фильтрации"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/typeproducts": {
            "get": {
                "tags": [
                    "TypeProduct"
                ],
                "summary": "Список",
                "operationId": "dab2a151b40dec6d4a65caa577003292",
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "TypeProduct"
                ],
                "summary": "Создание",
                "operationId": "af9395d6dbcb00d7d790fe943ef82566",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "product_name": {
                                        "type": "string",
                                        "example": "Дег 3л"
                                    },
                                    "quantity_produced": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Описание"
                                    },
                                    "status": {
                                        "type": "boolean",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/typeproducts/{typeproduct}": {
            "get": {
                "tags": [
                    "TypeProduct"
                ],
                "summary": "Список по ID",
                "operationId": "902ba22d74d855a9fab736b0df3b5f51",
                "parameters": [
                    {
                        "name": "typeproduct",
                        "in": "path",
                        "description": "ID тип продукта",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "TypeProduct"
                ],
                "summary": "Удаление",
                "operationId": "23bc17305f023014daf3580a12861875",
                "parameters": [
                    {
                        "name": "typeproduct",
                        "in": "path",
                        "description": "ID тип продукта",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "done"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "TypeProduct"
                ],
                "summary": "Обновление",
                "operationId": "dc5ccc2f8a4866b7b116214d6b5a6268",
                "parameters": [
                    {
                        "name": "typeproduct",
                        "in": "path",
                        "description": "ID тип продукта",
                        "required": true,
                        "example": 2
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "product_name": {
                                        "type": "string",
                                        "example": "Дег 3л"
                                    },
                                    "quantity_produced": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Описание"
                                    },
                                    "status": {
                                        "type": "boolean",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}